Changed stratum context call to login and added credentials
Fetch calls from React are now called with `credentials: 'include'` to
allow for cookies to be passed through the calls.

The proxied get context call to stratum is now calling login instead
since this is the new way of synchronizing cookie id with context. Hence
the name of the file and route is also changed accordingly